Moved EnumInteger to new package conv (conversion)
Moved EnumInteger to new package conv (conversion) as first step to refactoring the EnumInteger-System
Moved the appropiate Test
Updated imports in all accessors
