+++ src/main/java/dev/dokan/dokan_java/constants/microsoft/CreationDisposition.java
@@ -1,7 +1,7 @@
 package dev.dokan.dokan_java.constants.microsoft;
 
 import dev.dokan.dokan_java.DokanNativeMethods;
-import dev.dokan.dokan_java.constants.EnumInteger;
+import dev.dokan.dokan_java.conv.EnumInteger;
 import com.sun.jna.platform.win32.WinNT;
 import com.sun.jna.ptr.IntByReference;
