+++ src/main/java/dev/dokan/dokan_java/constants/microsoft/FileSystemFlag.java
@@ -1,7 +1,7 @@
 package dev.dokan.dokan_java.constants.microsoft;
 
 import dev.dokan.dokan_java.DokanOperations;
-import dev.dokan.dokan_java.constants.EnumInteger;
+import dev.dokan.dokan_java.conv.EnumInteger;
 import dev.dokan.dokan_java.structure.EnumIntegerSet;
 import com.sun.jna.platform.win32.WinNT;
